refactor: reusing latest python template setup for deploy
